requirements.txt
amqp==5.0.6
asgiref==3.3.4
billiard==3.6.4.0
celery==5.0.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.2.2
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-timezone-field==4.1.2
djangorestframework==3.12.4
importlib-metadata==4.0.1
kombu==5.0.2
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
python-crontab==2.4.1
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.1
